INO: Fix detection of variables to extract
When there were method arguments inside the block of an ifNilCheck:

m1: anArg
  ivar ifNil: [anArg doSomething]

They were not detected as variables to extract that had to be passed as
arguments to the new polymorphic selector. The problem was that
visitVariableNode does not visit the method arguments and that was the
method that MessageSendingBlocksVariablesToExtractFinder was using to
detect all variable nodes.