Stop adding trailing spaces to class autocompletion
There are several text fields in the ETMO and INO forms where the user
should enter a class name. Those inputs have autocompletion but after an
entry was selected a trailing space was added to the completed text. Now
it only completes with the entry without adding any separation
character.